[expo-router][refactor] split hooks.test.ios into multiple smaller test files
